Maintain target url path.
The UAA doesn't have to be run at the root of the web server.  For
example, in it might be http:://localhost:8080/uaa.

While a target url can be specified that includes a path, this path is
discarded when building API requests. For example, if the target is set
to http://localhost:8080/uaa, an info request won't result in a request
to http://localhost:8080/uaa/info but rather http://localhost:8080/info.

This commit changes this behavior to preserve the target url's path.