Refactor tests
Use gomega/ghttp for test assertions about requests. As more assertions
are made about the contents of token requests, ghttp will help with test
readability.

Additionally, because ghttp requires that every request to the test
server be handled individually, some interesting failure case behaviors
were uncovered.

Toekn retrieval is done as a side effect of calling Validate. Validate
is called multtiple times on construction of an API client. If an
invalid auth code is provided (or token retrieval fails for some other
reason) the token request is retried each time Validate is called. And,
because the default AuthStyle is used, each of these failed requests is
retried in hopes that providing the client credentials a different way
will result in success.

See here for more about AuthStyle

https://godoc.org/golang.org/x/oauth2#AuthStyle