Specifcy AuthStyle as AuthStyleInHeader
AuthStyle is a newer feature in the oauth2 library. This feature helps
the oauth implementation figure out how to pass along client credentials
to the UAA. These can be included in the form body or as a basic auth
header.

The go UAA client does not specify which AuthStyle it prefers. The
default is to try both to determine which method succeeds.

We know the UAA prefers the basic auth header, so set AuthStyleInHeader.