Improve client deserialization.
Prior to this commit, autoapprove field of a client was assumed to be
represented as an array of strings in JSON. However, it can be
represented as a boolean, a string, or an array of strings.

When autoapprove isn't an array of strings, deserialization fails.

This commit changes the Client deserialization logic, making the
autoapprove field an interface{} and allowing go to figure out the best
representation for the data. The field is now accessed via a function
which ensures that the value is always represented as a string slice.